refactor(commands): extract shared site and domain lookup helpers into a concern
Move safelyIterate, findSiteByName, and a 404-safe findPrimaryDomainId
into App\Commands\Concerns\FindsSiteResources so deploy and destroy stop
maintaining divergent copies of the same Forge lookup loops.
